# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for managing orders.",
        "title": "swagger Order Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/orderbook": {
            "get": {
                "description": "Retrieve the order book for a specific exchange and trading pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Get Order Book",
                "parameters": [
                    {
                        "description": "Order Book Request",
                        "name": "orderBookRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.OrderBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.OrderBook"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Save a new order book entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Save Order Book",
                "parameters": [
                    {
                        "description": "Order Books",
                        "name": "orderBooks",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.OrderBook"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orderhistory": {
            "get": {
                "description": "Retrieve the order history for a specific client.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Get Order History",
                "parameters": [
                    {
                        "description": "Client",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.HistoryOrder"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Save a new history order entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Save Order History",
                "parameters": [
                    {
                        "description": "History Order",
                        "name": "historyOrder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.HistoryOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Client": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "exchange_name": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "pair": {
                    "type": "string"
                }
            }
        },
        "entity.DepthOrder": {
            "type": "object",
            "properties": {
                "base_qty": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "entity.HistoryOrder": {
            "type": "object",
            "properties": {
                "algorithm_name_placed": {
                    "type": "string"
                },
                "base_qty": {
                    "type": "number"
                },
                "client_name": {
                    "type": "string"
                },
                "commission_quote_qty": {
                    "type": "number"
                },
                "exchange_name": {
                    "type": "string"
                },
                "highest_buy_prc": {
                    "type": "number"
                },
                "label": {
                    "type": "string"
                },
                "lowest_sell_prc": {
                    "type": "number"
                },
                "pair": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "side": {
                    "type": "string"
                },
                "time_placed": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "entity.OrderBook": {
            "type": "object",
            "properties": {
                "asks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DepthOrder"
                    }
                },
                "bids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DepthOrder"
                    }
                },
                "exchange": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pair": {
                    "type": "string"
                }
            }
        },
        "entity.OrderBookRequest": {
            "type": "object",
            "properties": {
                "exchange": {
                    "type": "string"
                },
                "pair": {
                    "type": "string"
                }
            }
        }
    }
}